pyproject.toml
[tool.poetry]
name = "emol"
version = "0.1.0"
description = "The Electronic Minister of the Lists"
authors = ["Ealdormere Web Team"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
authlib = "^1.3.0"
requests = "^2.31.0"
python-dateutil = "^2.9.0.post0"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
gunicorn = "^22.0.0"
boto3 = "^1.34.85"
mysqlclient = "^2.2.4"
django-permissions-policy = "^4.19.0"
django-csp = "^3.8"
django-dirtyfields = "^1.9.2"
django-markdownx = "^4.0.7"


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pylint = "^3.1.0"
types-requests = "^2.31.0.20240406"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
